+++ server/api/controllers/index.js
@@ -1,5 +1,6 @@
 'use strict';
 
+module.exports.user = require('./user');
 module.exports.container = require('./container');
 module.exports.image  = require('./image');
 module.exports.network  = require('./network');