+++ server/api/test/container.test.js
@@ -2,7 +2,7 @@
 
 let assert = require('assert');
 let request = require('supertest');
-let app = require('../../index');
+let app = require('../../../index');
 let chai = require('chai');
 
 process.env.NODE_ENV = 'dev';
@@ -26,11 +26,26 @@ let TEST_CONTAINER = {
   StdinOnce: false
 };
 
-describe('#container', function () {
+let login = {
+  name: 'admin',
+  password: 'admin'
+}
+
+let token;
+
+describe('#container', function() {
   var testContainer;
 
-  before(function (done) {
-    docker.createContainer(TEST_CONTAINER, function (err, container) {
+  before(function(done) {
+    request(app)
+      .post('/api/users/authenticate/')
+      .send(login)
+      .end(function(err, res) {
+        expect(res.status).to.be.equal(200);
+        token = res.body.token;
+      });
+
+    docker.createContainer(TEST_CONTAINER, function(err, container) {
       if (err) {
         done(err);
       }
@@ -44,6 +59,7 @@ describe('#container', function () {
     it('should list containers', (done) => {
       request(app)
         .get('/api/containers/all')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -55,6 +71,7 @@ describe('#container', function () {
     it('should list specific container', (done) => {
       request(app)
         .get('/api/containers/' + testContainer)
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -66,6 +83,7 @@ describe('#container', function () {
     it('should not list non-existent container', (done) => {
       request(app)
         .get('/api/containers/someMadeUpContainer')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           assert.equal(404, res.status, 'status should be 404');
@@ -79,6 +97,7 @@ describe('#container', function () {
     it('container should not be running', (done) => {
       request(app)
         .get('/api/containers/' + testContainer)
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -90,6 +109,7 @@ describe('#container', function () {
     it('container should start', (done) => {
       request(app)
         .post('/api/containers/' + testContainer + '/start')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -101,6 +121,7 @@ describe('#container', function () {
     it('should list running containers', (done) => {
       request(app)
         .get('/api/containers/running')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -112,6 +133,7 @@ describe('#container', function () {
     it('non existent container should not start', (done) => {
       request(app)
         .post('/api/containers/madeUpContainer/start')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(409);
@@ -125,6 +147,7 @@ describe('#container', function () {
     it('container should be running', (done) => {
       request(app)
         .get('/api/containers/' + testContainer)
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -136,6 +159,7 @@ describe('#container', function () {
     it('container should stop', (done) => {
       request(app)
         .post('/api/containers/' + testContainer + '/stop')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -147,6 +171,7 @@ describe('#container', function () {
     it('non existent container should not stop', (done) => {
       request(app)
         .post('/api/containers/madeUpContainer/stop')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(409);
@@ -160,6 +185,7 @@ describe('#container', function () {
     it('container should be stopped', (done) => {
       request(app)
         .get('/api/containers/' + testContainer)
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -171,6 +197,7 @@ describe('#container', function () {
     it('container should restart', (done) => {
       request(app)
         .post('/api/containers/' + testContainer + '/restart')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -182,6 +209,7 @@ describe('#container', function () {
     it('container should be started', (done) => {
       request(app)
         .get('/api/containers/' + testContainer)
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -193,6 +221,7 @@ describe('#container', function () {
     it('container should pause', (done) => {
       request(app)
         .post('/api/containers/' + testContainer + '/pause')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -204,6 +233,7 @@ describe('#container', function () {
     it('container should not restart', (done) => {
       request(app)
         .post('/api/containers/' + testContainer + '/restart')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(409);
@@ -215,6 +245,7 @@ describe('#container', function () {
     it('container should unpause', (done) => {
       request(app)
         .post('/api/containers/' + testContainer + '/unpause')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -226,6 +257,7 @@ describe('#container', function () {
     it('container should stop', (done) => {
       request(app)
         .post('/api/containers/' + testContainer + '/stop')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -236,10 +268,11 @@ describe('#container', function () {
   });
 
   describe('#pause', function() {
-    
+
     it('container should not pause', (done) => {
       request(app)
         .post('/api/containers/' + testContainer + '/pause')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(409);
@@ -247,10 +280,11 @@ describe('#container', function () {
           done();
         });
     });
-    
+
     it('container should start', (done) => {
       request(app)
         .post('/api/containers/' + testContainer + '/start')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -262,6 +296,7 @@ describe('#container', function () {
     it('container should pause', (done) => {
       request(app)
         .post('/api/containers/' + testContainer + '/pause')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -273,6 +308,7 @@ describe('#container', function () {
     it('container should be paused', (done) => {
       request(app)
         .get('/api/containers/' + testContainer)
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -287,6 +323,7 @@ describe('#container', function () {
     it('container should unpause', (done) => {
       request(app)
         .post('/api/containers/' + testContainer + '/unpause')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -298,6 +335,7 @@ describe('#container', function () {
     it('container should be running', (done) => {
       request(app)
         .get('/api/containers/' + testContainer)
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -305,10 +343,11 @@ describe('#container', function () {
           done();
         });
     });
-    
+
     it('container should not unpause', (done) => {
       request(app)
         .post('/api/containers/' + testContainer + '/unpause')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(409);
@@ -320,10 +359,11 @@ describe('#container', function () {
 
 
   describe('#remove', () => {
-    
+
     it('container should not be removed', (done) => {
       request(app)
         .delete('/api/containers/' + testContainer + '/remove')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(409);
@@ -331,10 +371,11 @@ describe('#container', function () {
           done();
         });
     });
-    
+
     it('container should stop', (done) => {
       request(app)
         .post('/api/containers/' + testContainer + '/stop')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -346,6 +387,7 @@ describe('#container', function () {
     it('container should be removed', (done) => {
       request(app)
         .delete('/api/containers/' + testContainer + '/remove')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -357,10 +399,11 @@ describe('#container', function () {
 
   describe('#create', () => {
     let container_id = '';
-    
+
     it('container should not be created without data', (done) => {
       request(app)
         .post('/api/containers/create')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(500);
@@ -372,6 +415,7 @@ describe('#container', function () {
     it('container should be created', (done) => {
       request(app)
         .post('/api/containers/create')
+        .set('x-access-token', token)
         .send(TEST_CONTAINER)
         .end(function(err, res) {
           expect(res.status).to.be.equal(201);
@@ -380,10 +424,11 @@ describe('#container', function () {
           done();
         });
     });
-    
+
     it('container should be removed', (done) => {
       request(app)
         .delete('/api/containers/' + container_id + '/remove')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -392,13 +437,14 @@ describe('#container', function () {
         });
     });
   })
-  
+
   describe('#stats', () => {
     let container_id = '';
-    
+
     it('stats should not be retrieved for non-existent container', (done) => {
       request(app)
         .get('/api/containers/madeUpContainer/stats')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(404);
@@ -410,6 +456,7 @@ describe('#container', function () {
     it('container should be created', (done) => {
       request(app)
         .post('/api/containers/create')
+        .set('x-access-token', token)
         .send(TEST_CONTAINER)
         .end(function(err, res) {
           expect(res.status).to.be.equal(201);
@@ -418,20 +465,22 @@ describe('#container', function () {
           done();
         });
     });
-    
+
     it('stats should be returned for container', (done) => {
       request(app)
         .get('/api/containers/' + container_id + '/stats')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
           done();
         });
     });
-    
+
     it('container should be removed', (done) => {
       request(app)
         .delete('/api/containers/' + container_id + '/remove')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);