+++ server/api/test/network.test.js
@@ -3,9 +3,9 @@
 var assert = require('assert'),
   request = require('supertest'),
   controllers = require('../controllers'),
-  app = require('../../index'),
+  app = require('../../../index'),
   chai = require('chai');
-  
+
 process.env.NODE_ENV = 'dev';
 
 const expect = chai.expect;
@@ -16,13 +16,32 @@ const docker = new Docker({
   socketPath: '/var/run/docker.sock'
 });
 
+let login = {
+  name: 'admin',
+  password: 'admin'
+}
+
+let token;
+
 describe('#network', () => {
-  
+
+  before(function(done) {
+    request(app)
+      .post('/api/users/authenticate/')
+      .send(login)
+      .end(function(err, res) {
+        expect(res.status).to.be.equal(200);
+        token = res.body.token;
+        done();
+      });
+  });
+
   describe('#create', () => {
-    
+
     it('should not create a network without a name', (done) => {
       request(app)
         .post('/api/networks/')
+        .set('x-access-token', token)
         .end(function(err, res) {
           expect(res.status).to.be.equal(500);
           done();
@@ -32,7 +51,10 @@ describe('#network', () => {
     it('should create a network', (done) => {
       request(app)
         .post('/api/networks/')
-        .send({Name: "testNetwork"})
+        .set('x-access-token', token)
+        .send({
+          Name: "testNetwork"
+        })
         .end(function(err, res) {
           expect(res.status).to.be.equal(201);
           done();
@@ -45,6 +67,7 @@ describe('#network', () => {
     it('should list networks', (done) => {
       request(app)
         .get('/api/networks/')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
@@ -55,10 +78,11 @@ describe('#network', () => {
     it('should list specific network', (done) => {
       request(app)
         .get('/api/networks/bridge')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);
-					expect(res.body.network.Name).to.be.equal('bridge');
+          expect(res.body.network.Name).to.be.equal('bridge');
           done();
         });
     });
@@ -66,6 +90,7 @@ describe('#network', () => {
     it('should not list non-existent network', (done) => {
       request(app)
         .get('/api/networks/madeUpNetwork')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(404);
@@ -73,11 +98,12 @@ describe('#network', () => {
         });
     });
   });
-  
+
   describe('#remove', () => {
     it('non-existent network should not be removed', (done) => {
       request(app)
         .delete('/api/networks/madeUpNetwork')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(409);
@@ -85,10 +111,11 @@ describe('#network', () => {
           done();
         });
     });
-    
+
     it('network should be removed', (done) => {
       request(app)
         .delete('/api/networks/testNetwork')
+        .set('x-access-token', token)
         .expect('Content-Type', /json/)
         .end(function(err, res) {
           expect(res.status).to.be.equal(200);